Исправляю недочёты ч.2: убрал всю логику из рест контроллеров.

+++ src/main/java/com/alex/warehouse/service/impl/AddressServiceImpl.java
@@ -2,9 +2,9 @@ package com.alex.warehouse.service.impl;
 
 import com.alex.warehouse.dao.BaseDAO;
 import com.alex.warehouse.entity.Address;
+import com.alex.warehouse.exception_handling.NoSuchDataException;
 import com.alex.warehouse.service.BaseService;
 import jakarta.transaction.Transactional;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,7 +21,11 @@ public class AddressServiceImpl implements BaseService<Address> {
     @Override
     @Transactional
     public List<Address> getAllEntity() {
-        return baseDAO.getAllEntity();
+        List<Address> addressList = baseDAO.getAllEntity();
+        if (addressList.size()==0){
+            throw new NoSuchDataException("Информация по данному запросу отсутсвует.");
+        }
+        return addressList;
     }
 
     @Override
@@ -33,12 +37,20 @@ public class AddressServiceImpl implements BaseService<Address> {
     @Override
     @Transactional
     public Address getEntity(int id) {
-        return baseDAO.getEntity(id);
+        Address address = baseDAO.getEntity(id);
+        if(address==null){
+            throw new NoSuchDataException("Адрес с id - " + id + " отсутствует.");
+        }
+        return address;
     }
 
     @Override
     @Transactional
     public void deleteEntity(int id) {
+        Address address = baseDAO.getEntity(id);
+        if (address == null) {
+            throw new NoSuchDataException("Адрес с id - " + id + " отсутствует.");
+        }
         baseDAO.deleteEntity(id);
     }
 }
