+++ src/main/java/com/alex/warehouse/service/impl/BlankServiceImpl.java
@@ -3,15 +3,12 @@ package com.alex.warehouse.service.impl;
 import com.alex.warehouse.dao.BaseDAO;
 import com.alex.warehouse.dto.BlankDTO;
 import com.alex.warehouse.dto.InvoiceDTO;
-import com.alex.warehouse.entity.Blank;
-import com.alex.warehouse.entity.Employee;
-import com.alex.warehouse.entity.Invoice;
+import com.alex.warehouse.entity.*;
 import com.alex.warehouse.exception_handling.NoSuchDataException;
 import com.alex.warehouse.mapping.BlankMap;
 import com.alex.warehouse.mapping.InvoiceMap;
 import com.alex.warehouse.service.BaseService;
 import jakarta.transaction.Transactional;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -21,23 +18,37 @@ import java.util.List;
 public class BlankServiceImpl implements BaseService<Blank> {
     private BaseDAO<Blank> baseDAO;
     private BaseDAO<Invoice> baseDAOInvoice;
+    private BaseDAO<Request> baseDAORequest;
     private BaseDAO<Employee> baseDAOEmployee;
 
-    public BlankServiceImpl(BaseDAO<Blank> baseDAO, BaseDAO<Invoice> baseDAOInvoice, BaseDAO<Employee> baseDAOEmployee) {
+    public BlankServiceImpl(BaseDAO<Blank> baseDAO, BaseDAO<Invoice> baseDAOInvoice, BaseDAO<Request> baseDAORequest, BaseDAO<Employee> baseDAOEmployee) {
         this.baseDAO = baseDAO;
         this.baseDAOInvoice = baseDAOInvoice;
+        this.baseDAORequest = baseDAORequest;
         this.baseDAOEmployee = baseDAOEmployee;
     }
 
     @Override
     @Transactional
     public List<Blank> getAllEntity() {
-        return baseDAO.getAllEntity();
+        List<Blank> blankList = baseDAO.getAllEntity();
+        if(blankList.size()==0){
+            throw new NoSuchDataException("Информация по данному запросу отсутсвует.");
+        }
+        return blankList;
     }
 
     @Override
     @Transactional
     public Blank saveEntity(Blank blank) {
+        if(blank.getId()==0){
+            throw new NoSuchDataException("Во входящих данных отсутствует id.");
+        }
+        Blank blankOld = baseDAO.getEntity(blank.getId());
+        if (blankOld == null) {
+            throw new NoSuchDataException("Котировка с id - " + blank.getId() + " отсутствует.");
+        }
+        blank.setDateCreate(blankOld.getDateCreate());
         return baseDAO.saveEntity(blank);
     }
 
@@ -83,12 +94,24 @@ public class BlankServiceImpl implements BaseService<Blank> {
     @Override
     @Transactional
     public Blank getEntity(int id) {
-        return baseDAO.getEntity(id);
+        Blank blank = baseDAO.getEntity(id);
+        if(blank==null){
+            throw new NoSuchDataException("Котировка с id - " + id + " отсутствует.");
+        }
+        return blank;
     }
 
     @Override
     @Transactional
     public void deleteEntity(int id) {
+        Blank blank = baseDAO.getEntity(id);
+        if(blank==null){
+            throw new NoSuchDataException("Котировка с id - " + id + " отсутствует.");
+        }
+        //т.к. бланк создаётся только приодобрении заявки, то при удалении его - заявка должна менять статус на "в ожидании"
+        Request request = blank.getRequest();
+        request.setStatus(new Status(2));
+        baseDAORequest.saveEntity(request);
         baseDAO.deleteEntity(id);
     }
 }
