+++ src/main/java/com/alex/warehouse/service/impl/InvoiceServiceImpl.java
@@ -3,13 +3,14 @@ package com.alex.warehouse.service.impl;
 import com.alex.warehouse.dao.BaseDAO;
 import com.alex.warehouse.dao.impl.WarehouseDAOImpl;
 import com.alex.warehouse.dto.InvoiceDTO;
+import com.alex.warehouse.entity.Blank;
 import com.alex.warehouse.entity.Invoice;
+import com.alex.warehouse.entity.Status;
 import com.alex.warehouse.entity.Warehouse;
 import com.alex.warehouse.exception_handling.NoSuchDataException;
 import com.alex.warehouse.mapping.InvoiceMap;
 import com.alex.warehouse.service.BaseService;
 import jakarta.transaction.Transactional;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -18,6 +19,7 @@ import java.util.List;
 @Service
 public class InvoiceServiceImpl implements BaseService<Invoice> {
     private BaseDAO<Invoice> baseDAO;
+    private BaseDAO<Blank> baseDAOBlank;
     private WarehouseDAOImpl warehouseDAO;
 
     public InvoiceServiceImpl(BaseDAO<Invoice> baseDAO, WarehouseDAOImpl warehouseDAO) {
@@ -28,7 +30,11 @@ public class InvoiceServiceImpl implements BaseService<Invoice> {
     @Override
     @Transactional
     public List<Invoice> getAllEntity() {
-        return baseDAO.getAllEntity();
+        List<Invoice> invoiceList = baseDAO.getAllEntity();
+        if(invoiceList.isEmpty()){
+            throw new NoSuchDataException("Информация по данному запросу отсутсвует.");
+        }
+        return invoiceList;
     }
 
     @Override
@@ -40,12 +46,24 @@ public class InvoiceServiceImpl implements BaseService<Invoice> {
     @Override
     @Transactional
     public Invoice getEntity(int id) {
+        Invoice invoice = baseDAO.getEntity(id);
+        if(invoice==null){
+            throw new NoSuchDataException("Накладная с id - " + id + " отсутствует.");
+        }
         return baseDAO.getEntity(id);
     }
 
     @Override
     @Transactional
     public void deleteEntity(int id) {
+        Invoice invoice = baseDAO.getEntity(id);
+        if(invoice==null){
+            throw new NoSuchDataException("Накладная с id - " + id + " отсутствует.");
+        }
+        //т.к. накладная создаётся только приодобрении котировки, то при удалении его - котировка должна менять статус на "в ожидании"
+        Blank blank = invoice.getBlank();
+        blank.setStatus(new Status(2));
+        baseDAOBlank.saveEntity(blank);
         baseDAO.deleteEntity(id);
     }
