+++ src/main/java/com/alex/warehouse/service/impl/StatusServiceImpl.java
@@ -2,9 +2,9 @@ package com.alex.warehouse.service.impl;
 
 import com.alex.warehouse.dao.BaseDAO;
 import com.alex.warehouse.entity.Status;
+import com.alex.warehouse.exception_handling.NoSuchDataException;
 import com.alex.warehouse.service.BaseService;
 import jakarta.transaction.Transactional;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,7 +20,11 @@ public class StatusServiceImpl implements BaseService<Status> {
     @Override
     @Transactional
     public List<Status> getAllEntity() {
-        return baseDAO.getAllEntity();
+        List<Status> statusList = baseDAO.getAllEntity();
+        if(statusList.isEmpty()){
+            throw new NoSuchDataException("Информация по данному запросу отсутсвует.");
+        }
+        return statusList;
     }
 
     @Override
@@ -32,12 +36,20 @@ public class StatusServiceImpl implements BaseService<Status> {
     @Override
     @Transactional
     public Status getEntity(int id) {
-        return baseDAO.getEntity(id);
+        Status status = baseDAO.getEntity(id);
+        if(status==null){
+            throw new NoSuchDataException("Статус с id - " + id + " отсутствует.");
+        }
+        return status;
     }
 
     @Override
     @Transactional
     public void deleteEntity(int id) {
+        Status status = baseDAO.getEntity(id);
+        if(status==null){
+            throw new NoSuchDataException("Статус с id - " + id + " отсутствует.");
+        }
         baseDAO.deleteEntity(id);
     }
 }
